MacroMeter/MacroMeter/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
tensorflow==2.14.0
opencv-python==4.8.1.78
numpy==1.26.2
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3 
google-generativeai==0.6.0
